# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "独角数卡 API V1",
        "description": "独角数卡项目的API V1版本，提供了RESTful风格的JSON API接口",
        "contact": {
            "name": "独角数卡开发团队",
            "url": "https://utf8.hk",
            "email": "Ashang@utf8.hk"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://unishop.wnecn.com",
            "description": "测试环境"
        }
    ],
    "paths": {
...
        "/api/v1/user/reset-password": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "重置密码",
                "description": "通过邮箱验证码重置用户密码",
                "operationId": "resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "重置密码示例",
                                    "value": {
                                        "email": "user@example.com",
                                        "code": "123456",
                                        "password": "newpassword",
                                        "password_confirmation": "newpassword"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "重置成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/otp/register": {
            "post": {
                "tags": [
                    "用户管理"
                ],
                "summary": "发送注册验证码邮件",
                "description": "向指定邮箱发送注册验证码，用于用户注册时的邮箱验证",
                "operationId": "sendRegisterEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterOtpRequest"
                            },
                            "examples": {
                                "example1": {
                                    "summary": "发送注册验证码示例",
                                    "value": {
                                        "email": "user@example.com"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "发送成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "required": [
                                                "data"
                                            ],
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/RegisterOtpData"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "$ref": "#/components/schemas/BaseResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseResponse": {
                "description": "通用响应结构基类",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "description": "状态码",
                        "type": "integer"
                    },
                    "message": {
                        "description": "响应信息",
                        "type": "string"
                    },
                    "data": {
                        "nullable": true,
                        "description": "响应数据"
                    }
                },
                "type": "object"
            },
...
            "RegisterOtpRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "description": "邮箱地址",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    }
                },
                "type": "object"
            },
            "RegisterOtpData": {
                "properties": {
                    "id": {
                        "description": "验证码ID",
                        "type": "string",
                        "example": "otp_123456"
                    }
                },
                "type": "object"
            },
...
            "ResetPasswordRequest": {
                "required": [
                    "email",
                    "code",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "email": {
                        "description": "邮箱地址",
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "code": {
                        "description": "邮箱验证码",
                        "type": "string",
                        "example": "123456"
                    },
                    "password": {
                        "description": "新密码",
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "description": "确认新密码",
                        "type": "string",
                        "example": "123456"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Bearer Token认证",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}